add state batch submitter unit test
